requirements/requirements.txt
Django==1.10.1
Faker==0.7.10
Pillow==3.3.1
django-guardian==1.4.6
factory-boy==2.8.1
mysqlclient==1.3.7
python-dateutil==2.6.0
six==1.10.0
